+++ app/src/plugins/prometheus/helpers.ts
@@ -101,6 +101,7 @@ export const jsonToProto = (json: any): Plugin.AsObject | undefined => {
 
   const plugin = new Plugin();
   plugin.setName(json.name);
+  plugin.setDisplayname(json.displayName ? json.displayName : '');
   plugin.setPrometheus(prometheus);
 
   return plugin.toObject();